PackageManager: Create missing userdata directory
..when app is moved to Adopted storage. Whenever an user tries to move
a package which is installed in more than one user to adopted storage,
it fails with the error ‘Not enough space’ even if
space is available on the adopted storage.
To fix the issue check that if the user data directory is present or
not. If the user data directory does not exist in the adopted storage,
attempt to create user data directories. This ensures that user
directories are created before the user data is copied from internal
storage to adopted storage. This prevents the move from failing.

Test: Perform the following steps:
1. Create multiple users apart from system user
2. Make ExternalLocTestApp and install for all the
   users.
3. Go to Settings > All Apps >  ExternalLocTestApp > Storage & Cache
4. Tap on Change under Storage User and Select Adopted storage.
5. Check that the process is not failing

Change-Id: I073f7f7686bd40ba1d126f04d125b0d2522190ee